Switch to using environment variables to enable/disable features
Use environment variables instead of using the custom feature
switch controller. This simplifies the code.

Added support for the MD_FEATURES_DISABLED environment variable. This
is now read on startup and used to configure the features.

Removed support for unspecified value for feature being enabled. The
behaviour here would be defined by the implementor of the feature
switch controller. To simplify this the addin only supports enabling
and disabling. The preferences UI has changed to support this and
now only shows a check box instead of a combo box for the enabled
states.